// src/app/modules/course/course.type.ts
import { Types } from 'mongoose'

type TDetails = {
  level: string
  description: string
}

type TTags = {
  name: string
  isDeleted: boolean
}

export type TCourse = {
  title: string
  instructor: string
  categoryId: Types.ObjectId
  price: number
  tags: TTags[]
  startDate: string
  endDate: string
  language: string
  provider: string
  durationInWeeks: number
  details: TDetails
}